Make the testspec more restrictive in terms of what can be set where.
Testspec is currently very permissive in very misleading ways.  In particular,
the tester parser itself will swallow K=V settings and apply them at the test
level, which breaks how a person would expect the scoping to work.  Other
settings apply to the entire simulation run globally, but appear to be workload
specific.  Even further, others affect simulation cluster creation or test
harness behavior, but can again be set anywhere in a testspec.

This changes testspec parsing to error if a setting that applies globally is
anywhere but the top of the file, or if a setting that applies test-wide is
applied to a workload instead of a test.