Add the special rangeLockVersionKey
This follows how metadataVersionKey is implemented in PR #1213. Specifically,
the GetReadVersionReply will return the range lock version that the client can
use to check if its local cache is valid.

This change only wires server side components. The next step is to create a
workload that will set the range lock version and check the version.