RemoveServersSafely:Safety check toKill1 to avoid cluster getting stuck
toKill1 and toKill2 are a random subset of all processes. If simply kill all processes in toKill1 or toKill2,
we may kill too many processes to make the cluster unavailable and stuck.

Similar as what toKill2 were modified if it can cause cluster unavailable,
we should do the same thing for toKill1