Add span info to mutations from proxy -> tlog
Updates serialization format of TLogCommitRequest.req messages sent from
the proxy to the transaction logs. Adds span context information for
each set of mutations so the transaction logs can identify the
transaction a mutation belongs to, allowing for better tracing of
requests through the system. This will soon be extended to the storage
servers as well, giving full visibility of the path of a transaction.