Randomize delay between locks using a poisson distribution
Also make the "delayBetweenLocks" parameter more accurate - i.e. if you
have a testDuration of 100 and a delayBetweenLocks of 1, you should
expect to lock about 100 times. If locking or unlocking takes
longer than delayBetweenLocks / 2 then you can't expect as many locks
and it will just kind of lock and unlock as fast as it can.