Add client side caches for range locks and versions
The range lock version is piggybacked in the GRV reply, just like the metadata
version key. The next step is to add a new RPC to retrieve the snapshot of
locks if the client doesn't have it. Then the client can check if get/set for
a key (range) is allowed.