Swap to -pthread from -lpthread
This is a (probably temporary?) fix to actually fix the compilation
issue with the added unit tests when compiling with gcc. The issue seems
to be that lpthread is getting linked before another library which
requires it, so linking fails. Linking pthread instead of lpthread fixes
the issue, but this may have an effect on other binaries in the system
as well.