fdbserver ignores range locks
The range lock is intended for clients. Using it on server side can block
transactions: each transaction needs to check if lock cache is stale; if so,
it tries to get the cache from commit proxies, which can cause timeout errors,
especially during recovery time. For instance, DD can be blocked.

Since fdbserver doesn't mutate data, it is safe to ignore range locks.