Make fdbmonitor understand $PID, so that --parentpid is usable on fdbserver.
Previously, --parentpid was only supported on windows.  #1496 extended that
to other platforms for a CI usecase.  On Linux, `prctl(PR_SET_PDEATHSIG,
SIGHUP)` should arrange for the subsequently exec()'d fdbserver to get a SIGHUP
and die if fdbmonitor is killed.  In local testing, I've confirmed that this
works, but I've also seen evidence that on some hosts/clusters it appears to
possibly not work.

Therefore, having --parentpid as a backup for the cases where it seems like
hosts might not properly support PR_SET_PDEATHSIG is nice.