Support commit multiple TLog groups with same commit version in TLogCommitRequest
A Tlog can belong to multiple Tlog groups, previously we submit
a TLogCommitRequest for each [TlogGroup, Tlog], thus a Tlog might
receive requests up to (#-of-Tlog groups it belongs to).

This commits allows TLogCommitRequest to contain multiple
Tlog group ids and prevVersions, together with the serialized
data to be persisted.

As a result, for each commit batch, each Tlog will receive at most
1 request.