Add fix for KillWorker unit test
The fix here is essentially to add additional work to the unit test. The
problem seems to be that during setup, some initial work is being done
with the key value store the ConfigNode opens. The simulator seems to be
doing work in the background to make the key value store "durable", but
because the unit test ends so quickly, the simulation is interrupted
(the key value files get deleted after the unit test is finished).