Persist data into TLog disk
TLog disk is used as spilling and restore, this change is about restore.

Previously data is only written to memory in partitioned txn code path,
it has to be written to disk as well for restore.

This commit also adds test by reading from an in-memory disk that is
injected to TLog.