Fix race between CF reads and CF writes
It was possible to read a CF at a version that does not yet have all of
the CF writes. This is problematic since the reader will only have a
subset of the actual mutations. To fix this, we only add a mutation
to the query's result set if it's version is less than the deque
version.