TLog pop works for on-disk data
Previously only in-memory data is popped, this commit makes
on-disk data work with pop API.

This PR also tests for spilling data, because when popping
we need to consider both spilling and disk queue.
Due to the way the code is written, I find no easy way to
disable spill while still popping, so they are tested together.

This commit also sets the spill threshold to default value
for tests who needs to peek, to avoid spilling because peek
does not work for on-disk data now, need fo fix it before
uncomment them.