To decide whether we need Standalone in versionMessages,there is a question: is it possible that versionMessages entry still exist, while the messageBlocks who has the corresponding Arena is already gone?
The answer is yes, because each time we call updatePersistentData,
a new version is passed in as a parameter and it is not the largest
version being erased -- instead, it is the first version not erased.

As a result, the first version not being erased could have its arena gone,
but the entry in versionMessages still exist, that would cause a
memory issue, thus we need to use standalone.

I understand this is causing unnecessary memcpy, but not sure if
there is a better way.