Fix static initialization order for maxAllowedVersion (#6570)
* Fix static initialization order for maxAllowedVersion

As it was defined, maxAllowedVerion was statically and globally initialized with a dependency on CLIENT_KNOBS. Initialization order was not spelled out and therefore this relied on chance specifics of order. In some cases these two were constructed in the wrong order and led to a segfault on startup. Fix by deferring initialization of maxAllowedVerion.

* Fix formatting with clang-format