Fix a rare race with version vector cache
The events that led to the assertion failure are:

Many txn started, some get ReadVersion1, some get ReadVersion2.
Recovery happened.
GRV proxy change detected and the client version cache is cleared.
Old GRV responded with ReadVersion1, and client version cache is updated to ReadVersion1.
A txn starts reading with ReadVersion2, triggering the assertion.

The fix is to throw an error to trigger transaction restart so that a new GRV
will update the version cache.