Fix TLogPeek duplicate versions read from disk
In corner cases, a version can be read from memory and then be
read from disk again, since we serialize data in disk first,
we need to de-dup when serialize data in memory.

When spill by value, each version as a key, and data as value.
When spill by reference, not each version is present as a key,
because multiple versions can be combined, in this case key
will be the last version