Optimization: support removing index conditionally
This change enables 2 modes:
* return all indexes
* return none of the indexes(except for boundary)

Note that boundary index entries need to be returned so
that caller can have a continuation, otherwise I do not see
an easy way to achieve this. Since we have serializable,
we might also need it to compare the set.

There will be a follow up change that enables more modes,
that return indexes based on whether there is a secondary query match.