mako: check for overflow in buckets
There is a downside that a bucket could overflow if there are too many latencies mapping to the same bucket. However, in the vast majority of situations this is unlikely to occur. While mako can be run with an unbounded amount of TPS, our tests have seen a maximum of 954013 TPS. Then running mako for 2 hours (3600 seconds) we obtain (3600 * 954013) / 10 = 343444680 as a bucket count with a sampling rate of 10. Since each bucket is stored in a uint32_t, the maximum is near 4 billion and this example doesn't bring us close to that range. Hence overflowing while increasing the sample rate is an extremely rare scenario.