mako: Always add OP_COMMIT and OP_TRANSACTION to map
This commit makes sure that we always create OP_COMMIT and OP_TRANSACTION entries in the DDSketch map. The reason is because even though they are explictly set as active operations through the "txn" string, it is implctly used in stats collection.

There was also a bug in ThreadStatistics::mean() where it was incorrectly reporting the max instead.