Optimization: Optimistically tag and serialize mutations pre-resolution.
Tagging and serializing mutations currently happens post-resolution,
after a commit version has been assigned. This can result in head-of-line
blocking by large transactions, as they must complete this CPU-intensive
process prior to evaluating any later transactions.

This optimization attempts to tag and serialize mutations pre-resolution,
prior to getting a commit version, on the assumption that the key->tag
mapping changes infrequently.

We add a `preResolution` tag to `assignMutationsToStorageServers` so that
function can be called before resolution. If we detect that any metadata
updates have taken place between the time we tagged the mutations, and
the time we would have tagged them post-resolution, then we throw away
the tagged and serialized mutations and regenerate them.

These are the current set of conditions which invalidate this optimization:

* any metadata mutation, whether it went through this commit proxy, or we
  found out about it from the resolver

* any transaction in the batch doesn't commit; we have no way to disentangle
  individual transactions' mutations once serialized

* resolver private mutations are enabled; these will assert that they are
  the first mutations added to `toCommit`, which is not possible if we've
  already pre-tagged the mutations

* encryption is enabled; cipher keys aren't available until `getResolution`
  so we can't serialize anything beforehand

* any versionstamp vorrection; these mutations are modified prior to sending
  them to resolver in a way that they no longer look like metadata updates
  afterwards.